Let me resume. I've been thinking during the night.
I've taken the problem description for its word that making pal be `l = rev l` is a bad idea, but I really haven't been getting much traction with anything else.

Today I want to play with the palindrome exercises a little. After that, I do not think there are any 4 or 5 star exercises remaining and will be able to mop up the chapter.

I've still yet to study that remember example too.

So let me get this out of the way. It is time to slowly bring the very long inductive propositions chapter to a close. It took me an insane amount of time already.