Like yesterday, I've entered a cycle of perplexity. It is different than the last time I got stuck on the pumping lemma. The pumping lemma was complicated - I did not really understand what I was proving there, so I could only rely on my technical skills.
But this is different. Palindromes are not exactly a hard concept to grasp and neither is what the theorems are saying. I pretty much understand everything that is needed...or at least that should be the case.

Every time I stop to prove, it is just like the intuition gets cut off all of a sudden. It is like stepping into a lighted room only for it to go dark every time I step over the threshold.

Back in 2016 this really annoyed me about Coq and made me decide to skip dependent types. But now that I am interested in it specifically for math, I find that this sort of situation is absolutely splendid. It is exactly these sorts of scenarios that provide the ideal training experience.

Problems like the pumping lemma only train the surface level of what it means to prove things.

Overcoming problems like the one I am facing now will change the nature of my talent. If I can overcome whatever mental errors are causing my mind to short circuit, it will enable a different kind of thinking. I absolutely won't skip or try to rush through this problem. I'll dig in and come out the other way.

Let me step away from the screen. I'll dwell on this while I take a nap.