"11:40am. https://stackoverflow.com/questions/56074850/how-to-prove-that-given-an-equality-that-a-string-is-its-reverse-its-tail-is-al
I am really curious as to what I will get in response.

11:45am. While I wait, let me work on something else.

```
Theorem rev_split' :
  forall X (l : list X),
  l = rev l ->
  (exists k, l = k ++ rev k) \/ (exists k x, l = k ++ [x] ++ rev k).
```

It is a bit similar to this, except I want to be able to split lists into two equal halves. This was a thing in Idris.

```
Theorem list_halve :
  forall X (l : list X),
  (exists a b, l = a ++ b /\ length a = length b)
  \/ (exists a b x, l = a ++ [x] ++ b /\ length a = length b).
```

This. Let me take a break here. I'll reread the Idris book after I get back.

It seems I will be stuck on this 5-star problem for a while yet."