I really was in thought for quite a while it seems. I've decided. Though the `palindrome_converse` problem on the surface seems like it is about palindromes, the main challenge in it is about splitting lists. Once I am able to do that, things will get much easier for me. It is not easy right now.
I've put in quite a bit of time, and I am still not sure how to do the inductive proposition for halving lists.

The reason why I am having this much trouble with these palindrome proofs is because my understanding on how to construct the propositions is still so lacking.

```
Inductive pal {X : Type} : list X -> Prop :=
  | pal_nil : pal []
  | pal_fst : forall x l, pal (rev l ++ l) -> pal (rev l ++ [x] ++ l)
  | pal_snd : forall x l, pal (rev l ++ [x] ++ l) -> pal (rev l ++ [x; x] ++ l).
```

I've made this, but how do I go from say `pal (rev l ++ [x; x] ++ l)` to `pal l`? This is a complete mystery to me.

I clearly have missed a few steps since that is the case.

Tomorrow I will play around a little. Rather than trying to halve lists right away, instead I will replicate some things from the Idris book. Like for example a view that takes from the end of a list and such.

I've put in a lot of work into this chapter, but clearly I haven't really internalized the essence of it. When I am done with it, that will no longer be the case.