"1:05pm. Well that was fun. Let me resume work though.
Where was I? Ah, yes, let me do the induction on `b`.

1:20pm.

```
Theorem disjoint_iff_nodup_app :
  forall {X} (a b : list X),
  disjoint a b <-> NoDup (a ++ b).
Proof.
unfold disjoint, not.
split.
- intros.
  generalize dependent a.
  induction b; intros; simpl.
  + induction a.
    * constructor.
    *
```

```
1 subgoal
X : Type
x : X
a : list X
H : forall x0 : X, In x0 (x :: a) /\ (In x0 [ ] -> False)
IHa : (forall x : X, In x a /\ (In x [ ] -> False)) -> NoDup (a ++ [ ])
______________________________________(1/1)
NoDup ((x :: a) ++ [ ])
```

Right now I am stuck on the first branch of this.

1:35pm. Stop surfing /a/ me.

1:40pm. Wait, wait. I've neglected something really important. `True -> False` and `False -> False` cannot both be false.

```
disjoint a [] -> NoDup (a ++ []).
```

Like suppose I have this. What exactly should I expect?

This essentially simplifies to `In x a -> NoDup a`. That cannot be right. The premise certainly does not imply the conclusion. I am a fool for trying to prove this.

1:45pm. `NoDup a -> In x a` would certainly be implied.

Well the exercise did say that I only need to prove one thing, so

```
NoDup (a ++ b) -> disjoint a b
```

This is certainly provable. But let me make a setup where `disjoint` is on the left side as an exercise. I am doing this whole thing as a training run anyway.

1:50pm. I am a bit dry.

```
disjoint [a] [b] -> NoDup (a ++ b)
```

I suppose this qualifies.

Yeah, I have nothing particularly interesting. Let me just prove the first one and I will move on. There will be plenty of exercises after this so it not like I will run out.

1:50pm. Let me step away from the screen for a bit. I should do the chores now."