"I am stuck hard on this problem. I am completely boxed in.
```
(** First prove an easy useful lemma. *)

Lemma in_split : forall (X:Type) (x:X) (l:list X),
  In x l ->
  exists l1 l2, l = l1 ++ x :: l2.
```

No seriously, screw this. Why would this be useful? The same goes for the excluded middle making things "easier".

I hate the useless cryptic comments while I am struggling with a problem. Let me take a break here.

I'll do the chores and then sleep on this for a bit. Just heating my brain is not going to do much at this point.

Like the pumping lemma and those before it, I am going to be stuck on this for a while. It was a miracle that I even managed to do some of those, so it is not like there is a guarantee that I am going to be able to do this one."