"6:20pm. https://youtu.be/gJenf1DJfBQ?t=493
No I cannot follow this. It might as well be in some other language I do not understand.

http://www.math.lsa.umich.edu/~ablass/kfin.pdf
"An induction principle and pidgeonhole principles for k-finite sets"

6:25pm. https://github.com/llbit/Pigeonhole-Principle

Here is the solution to the thing.

https://github.com/llbit/Pigeonhole-Principle/blob/master/Pigeonhole.pdf

Let me read this.

"The Software Foundations exercise first suggests proving the following lemma. It may not
be clear for a long time why this lemma is useful, but I could not prove pigeonhole principle
without it."

```
Lemma in app split : ∀ T x (a b : list T),
In x (a ++ b) ↔
In x a ∨ In x b.
```

For a bit I thought this was it, but not really.

6:50pm. Done with lunch. Let me go through this paper so I can put my heart at ease and then I will call it a day.

...He defined the thing as I did originally.

6:55pm. Ah, I did not see the transitivity relation.

7pm. There is a lot of stuff in the paper. I will go through it properly tomorrow. I do not recognize a lot of the things being used in it. I should also figure out how to do the proof using the excluded middle.

I have no idea why this problem is only 4 stars. It should be 5 stars or more at least. Hopefully I'll learn something from this.

My personal principle is that a few hours of work per a single problem should be enough before looking at the answer.

One mistake I really should have avoided is to rewrite `In x a -> In x b` to `AllIn a b`. That might have made things clearer.

Well, let me rest here. I will get this done tomorrow and then finally put the close on this very long chapter.

The stuff here is beyond crazy. I will be doing this for the next six months at this rate."