Let me chill for a bit and then I will...damn it is cold here. Summer is nearly upon us, but lately the weather is more like winter. I guess I will be chilling literally until the room heats up.
When I got Clusterfuck Doom I did not have much expectation of it, but as it turns out the thing is really fun. I am in the deeply enjoyable obsessive stage with it - it was the same with Brutal Doom and Demonsteele when I first got them. The main point of Clusterfuck are the boss battles. The legendaries are just giant balls of death - when I ran into the Nemesis Cyberdemon yesterday it started screaming that it is a god while making the level shake with every attack. That was damn awesome.

Too bad cover cannot be damaged. Once I get to the point where I can make Hell, those kinds of battles will take entirely new dimension.

Well, let me read Hino for a bit and then I will start this thing. If I want my power to get to the point where I am able to make the Games that I want, I require hard training.