Now that I am done with the break I think I will step away from the screen for a while just to think. I feel like I need a redo. I thought I was making good progress through the book, but I am not. I do not feel like messing the exercise I am doing at the moment.
All the programs that I've done in the past had a certain stickiness to them. I can clearly visualize them in my mind, but the Coq proofs are not like that. Even if it is something simple, but not trivial like `m * n = n * m` I cannot completely recall it. More worryingly I cannot recall the high level reasoning behind the proof either. Why means I haven't learned the most essential parts of programming in this new domains in the past 3 weeks.

It is all a jumble of random rewrites. I could hack it dues to my programming experience, but if I want to really move on to the next level, I need to gain a deeper understanding of this. I need to internalize every single thing that I do just like with vanilla programming.

I think that the pigeonhole proof in particular is 2 levels beyond where I am at now.