"I ended up helping around the house and spending all this time in bed thinking. I am still in thought. Ultimately, it is better that I rebuild my motivation on a firmer footing than mess around.
I am just lacking too much. I haven't really internalized the techniques in the past chapters properly which is why I am having so much difficulty in the pidgeonhole problem. It won't be the only one either. If it was not this one, the same would have happened at some point.

Rather than focusing on overcoming depression, it is actually better to forget the old reasons and rebuild them from the ground up so they point towards the same goal. I've sort of been hoping that I would grow out of my beginner status with enough effort, but in some ways the whole process is flawed. The effort will be necessary of course, but I need to go at it with a different midset.

I've gotten stuck in the past and every time overcome it - always I felt a sense of relief after that, but I haven't faced up to it. I've gotten stuck in vanilla programming as well and overcome it with effort, but this is different. This cannot possibly so difficult. I cannot only rely on my improved technical skills to carry me. If I have a strong intuition that a proof is doable then I should be able to trace the process 100%.

If I cannot then there is something wrong with my whole mental machinery.

4:15pm. Well, let me dwell on this for a while longer (while I read Kengan Omega), and I will call it a day after 5pm. Tomorrow I will try dealing with the 3-star exercise I am doing now.

If it was the pre-pigeonhole me, this would already been done, but rather than going at it with a fiery passion, I want to try going at it full of doubt - at both myself and the problem. I've been relying on Coq way too much to prove things and it is really costing me.

What a need to attain is a more penetrating insight. I could get that if I overcome this, but I must not do it the same way as I have for past problems. I have to endure the darkness, until I find the thread that I am seeking. I need to overcome my habit retreating back into the light when it gets tough.

I should know that there is absolutely nothing of value there."