"4:40pm. Let me see if I can clear out `H : s1 =~ re1` a few steps back and still do the base case.
No, it has to be there. Absolutely.

4:50pm. No, it cannot be done. Probably I made a mistake inducing on `s1`, rather than on `H` itself.

4:55pm.

```
    * simpl.
      inversion x0.
      subst.
```
```
1 subgoal
re2 : reg_exp
s : list ascii
a : ascii
x0 : [a] ++ s = a :: s
IHexp_match2 : forall (a : ascii) (s0 : list ascii),
               s = a :: s0 -> s0 =~ derive a re2
H0 : s =~ re2
______________________________________(1/1)
s =~ App EmptySet (derive a re2)
```

No seriously....

No way can I construct the `EmptySet`.

5pm. Yeah, I hate Coq. How long have I been doing the `IndProp` chapter by now?

I am going to finish it even if it kills me, but if the next chapter turns out to be more of this, I am seriously going to consider skipping the rest of the SF and just move on to studying the math library. I just do not have the time for this garbage.

I might be stuck on another problem right now, but at this point I just do not care. If I cannot do this tomorrow, I will move to studying Agda for a while so I can figure out the pigeonhole principle proof.

5:10pm.

```
1 subgoal
s1 : list ascii
re1 : reg_exp
s2 : list ascii
re2 : reg_exp
H : s1 =~ re1
H0 : s2 =~ re2
IHexp_match1 : forall (a : ascii) (s : list ascii),
               s1 = a :: s -> s =~ derive a re1
IHexp_match2 : forall (a : ascii) (s : list ascii),
               s2 = a :: s -> s =~ derive a re2
a : ascii
s : list ascii
x : s1 ++ s2 = a :: s
______________________________________(1/1)
s =~
Union (App (derive a re1) re2)
  (App (if match_eps re1 then EmptyStr else EmptySet) (derive a re2))
```

I have no idea what to do here. The difficulty setting for this game is just too damn high.

I am done for the day here.

It is time for some play.

I've definitely entered a new era of strife. To think I will be fighting this beast for the next few years. Win or lose, if somebody finds these commits and reads them, they will bear witness to my iron will.

I do not know whether there is a point in working or doing these exercises. In a just world, I would be able to enter the self improvement loop as soon as I realized the true essence of it.

While I might be doing all this work, it is not like I necessarily believe hard work is a virtue. Rather, I believe in the desire. I believe that ultimately the universe should bend to desire and I am willing to pay whatever price is necessary to attain such a goal. This is a hard thing to achieve, but by focusing on the contradictions in such an ideology, a way forward will be found. Ultimately, any ideology has its flaws, but a strong one can accept them and surmount them.

Those currently floating about in the consensus consciousness will not make the cut. The left wing cannot look at reality, and the right is just its shadow and cannot turn despair into power. The right could eventually become its own thing, but then at that point it would not be the right anymore. I haven't decided yet whether in either case the problem is the ideology or whether its hosts are just too low spec.

Well, it does not matter. What matters is what I think. What matters is that I pay the price. That is the most I can do as a human.

This trash work of cleaning up the proofs is a part of that. I will not shirk away from this duty.

If I do not feel like doing these proofs I will study Agda or the libraries or something else related to this. Eventually I will get over my blindness and gain the ability to think more in 'should's. I will believe that if I keep practicing, my intuition will eventually improve."