I've chilled for a while, but it is time to start. I really slept quite well tonight, so I am quite fresh at the moment. Let me see if I can put a dent into that exercise.
If only I could do that, I will be able to finish the chapter today. Nearly 3 weeks already. Holy shit.

That is for how long I've been working on it.

The thing about skill is - one cannot beg or hope for it. One just needs to go out there and get it with his own hands.

Today will be another day in a long line of days of me doing that.