Done with lunch. And I am done for the day. It honestly does not feel like I did much today. And indeed, it seems I did some theorem proving in the morning and then got pissed off. Then I spent a lot of my time either getting Agda to work or getting distracted by unrelated work.
But that is how things are somedays. These last exercises in the IndProp chapter are just so insanely difficult and I am out of drive to keep plowing through relentlessly for the time being. I feel like I am just missing experience, but just pushing harder and harder is not a good way to become better at this. I found where my limits lie. If I had to do anything more on that level, I'd rather study libraries and come back to this later when I am better armed to deal with this.

Now let me have some fun. I've earned it.