I've chilled for a bit and now it is time to start. I am going to get through started chapters of the PLF in Agda book. This stuff is easy so I should not have to spend too much time on it.
After the bloodletting that were the last 3 weeks, I need to take it easy for a while. I'll rebuild from the top and then do these problems properly with my full power.

I know it sounds like an excuse that I can't hack it in Coq, but the language being unwieldy will in fact impact me in detrimental ways.

The foremost problem is that a language that is poor will also stick poorly in memory. This is one of the issues I have with past proofs in Coq right now. It is not the case for me in either F# or Spiral - I can always remember what everything does there. It surprised me when I did Idris that some of the proofs had a lot of structure in common with regular programs.

If suddenly you bring in backwards reasoning and replace good language design with tactics, that gets obscured to a major degree.

Of course, in hindsight it should have made sense that proofs themselves should have connection to standard functional programs, but when I was doing Coq in 2016 I had not realized that at all.

Once I am done with the Agda book...

It might be worth to going through Logical Foundations (the first SF volume) and do it in Agda instead of Coq.

What I've written sounds nice in theory, but I should really test whether I am getting concrete benefits by doing functional programming rather than tactics-based proving.

This is quite a lot of work and I am already tired of it, but that was the case back in 2015. As long as my flame burns, I will make a single step forward.

I guess it was inevitable that it would come to this long trek through the desert. Deep down I suspected that with unsound foundations, I will not be able to attain what I desire. I need to build them. Then the algorithms will come to me, and so will everything else.