"10:10am.
```
+-identityʳ : ∀ (m : ℕ) → m + zero ≡ m
+-identityʳ zero =
  begin
    zero + zero
  ≡⟨⟩
    {! 2  !}
  ∎
```

First thing's first, let me study how these equation chains work. I am guessing it just holds one side constant and operates on the other.

10:20am.

```
+-identityʳ : ∀ (m : ℕ) → m + zero ≡ m
+-identityʳ zero =
  begin
    zero + zero
  ≡⟨⟩
    zero
  ∎
+-identityʳ (suc m) =
  begin
    suc m + zero
  ≡⟨⟩
    suc (m + zero)
  ≡⟨ cong suc (+-identityʳ m) ⟩
    suc m
  ∎
```

I just realized while playing with holes here that `C-c C-space` will actually typecheck them. That is great, that will make things much easier. `≡⟨ cong suc (+-identityʳ m) ⟩` the thing inside the brackets here affects the right side of the equations.

```
  ≡⟨⟩
    suc (m + zero)
```

The thing on the outside of the `⟨⟩` affects the left side, in fact I think it is the entire construction. I get it.

I could probably...

```
  ≡⟨ (+-identityʳ m) ⟩
    ?
```

No this gives a type error.

10:25am. Ok, I get it. Let me move on.

10:35am.

```
+-comm : ∀ (m n : ℕ) → m + n ≡ n + m
+-comm m zero =
  begin
    m + zero
  ≡⟨ +-identityʳ m ⟩
    m
  ≡⟨⟩
    zero + m
  ∎
+-comm m (suc n) =
  begin
    m + suc n
  ≡⟨ +-suc m n ⟩
    suc (m + n)
  ≡⟨ cong suc (+-comm m n) ⟩
    suc (n + m)
  ≡⟨⟩
    suc n + m
  ∎
```

This one is a bit more interesting. I wonder if `≡⟨ +-suc m n ⟩` and `≡⟨ cong suc (+-comm m n) ⟩` compose with each other? I'd expect that to be the case.

10:45am. The things inside the brackets do rewriting so yeah, in fact they do commute.

```
  ≡⟨ +-suc m n ⟩
    ?
```

The holes can be filled in automatically using `Solve Constraints` or `C-c C-s`.

11:10am.

```
+-comm′ : ∀ (m n : ℕ) → m + n ≡ n + m
+-comm′ m zero rewrite +-identity′ m = refl
+-comm′ m (suc n) rewrite +-suc′ m n | +-comm′ m n = refl
```

So this is how it is possible to chain the rewrites. I see.

11:25am.

```
+-swap : ∀ (m n p : ℕ) -> m + (n + p) ≡ n + (m + p)
```

Let me do the first exercise and then I will take a break.

```
+-swap : ∀ (m n p : ℕ) -> m + (n + p) ≡ n + (m + p)
+-swap m n p rewrite sym (+-assoc m n p) | +-comm m n | +-assoc n m p = refl
```

This has a lot of stickiness for me. I can easily internalize this. In fact this is similar to how I did the same proofs in Coq. Unfortunately, SF introduces how to do function application much later, so on my first run back in 2016 I had to do without it. The end result was much uglier.

Agda has a bit awkwardness in that unlike Coq it needs constant recompilation when doing the rewrites, but I can deal with that.

Let me stop here. The stuff that comes after this are bog standard laws on naturals. They should make a decent warm-up for what is to come."