Let me start the day. I seem to be having too much fun Clusterfuck after the day is done, but I've slept well tonight.
Today I will go through the chapter I started yesterday from the top. I think I will see if I can get the literal Agda files to work. I've noticed that the files do not get loaded if another is already opened.

At first I assumed that literal files are not supported with plugin, but then realized that normal files do not get loaded if another is in progress in Atom. And every time I've tried it I had something else open. So trying this out will be my first task.

The plugin is quite buggy so I might open an issue for it not highlighting and locating the errors correctly.

Well, let me start getting into it.

I think the Agda book is off to a good start - it explains everything and does not insert random ultra high difficulty exercises on what should be an introductory book.

I like the language's elegance and the vigor resulting from that. The code I am writing here definitely feels like a work of art, but it remains to be seen if this will hold after I start doing the harder exercises such as those that are making me stumble in Coq.

Only one way to find out.