"9:35am. The literal Agda file works fine. Great.
9:45am. Ah, I just had to read that OpenAI article.

https://techcrunch.com/2019/05/18/sam-altmans-leap-of-faith/

I am talking about this one.

"Continued Altman, “We’ve made a soft promise to investors that, ‘Once we build a generally intelligent system, that basically we will ask it to figure out a way to make an investment return for you.'” When the crowd erupted with laughter (it wasn’t immediately obvious that he was serious), Altman himself offered that it sounds like an episode of “Silicon Valley,” but he added, “You can laugh. It’s all right. But it really is what I actually believe.”"

How shameless.

> When the crowd erupted with laughter (it wasn’t immediately obvious that he was serious)

I am kind of miffed at OpenAI for trying to embrace mad science - it feels that they are trying to encroach upon my territory. I know that I am laughably trying to cause the apocalypse, but I feel that an organization that would try and to do this seriously should have a different set of values than OpenAI does. They lack the self awareness needed to cause the Singularity. If they want to enter the self improvement loop, they should signal some awareness of what it actually is.

Just what do they intend to do with those huge computers? Do they not see that they need better algorithms instead? Can they not adjust their initial optimism to a more realistic measure as they go along?

Surely they must, so I am in awe at the size of the scam they are trying to run.

Billions of dollars raised? Just what should the investment return on taking the universe apart be? Imagine just how stupid it would be for an AI capable of operating at that scale to focus on making dollars for its backers.

The bar where making money stops making sense is actually much lower than that - simply having better credit assignment algorithms would do the trick.

10:05am. I am getting really distracted by this. Let me get back into the book.

```
#### Exercise `orderings` {#orderings}

Give an example of a preorder that is not a partial order.

\begin{code}
-- Your code goes here
\end{code}

Give an example of a partial order that is not a total order.

\begin{code}
-- Your code goes here
\end{code}
```

My brain is a bit asleep so I do not feel like doing this.

...Have I ever run into an ordering that was not total? I cannot recall anything.

10:20am.

```
≤-total : ∀ (m n : ℕ) → Total m n
≤-total zero    n                         =  forward z≤n
≤-total (suc m) zero                      =  flipped z≤n
≤-total (suc m) (suc n) with ≤-total m n
...                        | forward m≤n  =  forward (s≤s m≤n)
...                        | flipped n≤m  =  flipped (s≤s n≤m)
```

Let me see if the editor will help me destructure the view like in idris.

Nope, I had to arrange it by hand.

When I run into a highlighting error again, I will open an issue along with a screenshot.

10:30am.

```
#### Exercise `*-mono-≤` (stretch)

Show that multiplication is monotonic with regard to inequality.

\begin{code}
-- Your code goes here
\end{code}
```

Finally comes the first exercise. I will eat this thing.

10:35am.

```
open import Data.Nat using (_*_)
open import Data.Nat.Properties using (*-comm)

*-mono-≤ : ∀ (m n p q : ℕ)
  → m ≤ n
  → p ≤ q
    -------------
  → m * p ≤ n * q
*-mono-≤ m n p q m≤n p≤q  = {!   !}
```

I am going to use the same structure as in the proof of the +. To do that I need to prove the two sides.

10:50am.

```
≤-cong : ∀ (m n : ℕ) → ∀ (f : ℕ → ℕ) → m ≤ n -> f m ≤ f n
≤-cong .0 n f z≤n = {! z≤n {n = f zero}  !}
≤-cong .(suc _) .(suc _) f (s≤s m≤n) = {!   !}
```

I am on the wrong track here. Let me instead use the + monotinicity.

11am.

```
≤-cong : ∀ (m n : ℕ) → ∀ (f : ℕ → ℕ) → m ≤ n → f m → f n → f m ≤ f n
≤-cong m n f fm fn m≤n = ?
```

Holy shit the editor issues are making this quite harder than it should be. Every time I compile it points to an error 20 lines up.

```
E:/TheoremProving/plfa/Relations.lagda:543,48-51
ℕ should be a sort, but it isn't
when checking that the inferred type of an application
  ℕ
matches the expected type
  _124
```

Just what is wrong with...

11:10am. I've decided to open an issue on the plugin repo. I can't bear this anymore. I am going to have to go with Emacs after all. That should give me a stable experience once I get it running.

Getting familiar with Emacs was not on my TODO list, but I might as well put it there.

I'll take a break from the book for the time being until I get this done. Let me take an actual break here since it is 11am."