I chilled for a bit and now it is time to start the day.
What I did before 2019 was really directed in terms of intention. The plan was to make Spiral and then just crush poker with the right algorithm that was surely out there. The plan was to encircle the enemy before moving in for the attack.

The right algorithm not being in any of the papers makes things a lot more complicated for me. There is nothing I can invent and piece together to just make what I want.

All of my programming skill is useless in this situation. All the advantages are held by mathematicians and I am not one.

But since I have time, I might as well make use of it instead of sucking my thumb. Maybe I am set as far as programming is concerned, but my limits in math have scarcely even begun to be explored.

So for my remaining time on this Earth, that is what I will be doing. I will forget about being an inventor and just plunge into the darkness in search of the key to my future.

It is really lame. Having a purpose and following it for the past 15 years has not gotten me much. All of my gains are intangible and all my failures are stark. It is just so unfair that despite realizing the true nature of the self improvement loop and accepting it, I cannot even come close to the process. If my luck turns out to be bad, I might not even live to see myself enter the loop.

Since this is the domain of discovery, I cannot count on steady progress to get me where I want like when I was making Spiral.

And as uncertain as me finding the right algorithm in the first place is, it would have to be something absolutely phenomenal to make a new wave of machine learning.

The odds are so stacked against me in this that if I can do this, I am going to forgive myself for all the past failures and forget all my regrets.

I will begin to see myself as something new.

Forget money. Forget hedging my bets with side activities like trying to beat poker and making a language.

From here on out, I will do what I should have done since the very first day. I should have attained these skills from the time I utered my first word.

I need to grow up and just do this. From the beginning I knew that all the real power is held by those with the capability to create algorithms. It is time to go get it.

Let me start the Equality chapter.