"1:05pm.
```
  +-monoʳ-≤ (suc n) p q p≤q  =
    ≤-begin
      (p ≤ q)
    ≤⟨ s≤s (+-monoʳ-≤ n p q p≤q) ⟩
      (n + p ≤ n + q)
    ≤∎
```

Honestly, I am not sure how to deal with this.

Let me try something.

```
  ≤-begin_ : ∀ {x y : ℕ}
    → (a : Set)
    → {a}
      -----
    → a
```

I need to go in this direction, but if I make it like this, I lose the information that I am dealing with inequality here.

1:10pm.

```
  ≤-begin_ : ∀ {x y : ℕ}
    → (a : Set)
    → {a = x ≤ y}
      -----
    → a
```

This might be promising. I can't really think of much better than going in this direction.

Let me try it.

1:15pm.

```
  ≤-begin_ : ∀ {x y : ℕ}
    → {a : Set}
    → {a = x ≤ y}
    → a
      -----
    → a
```

No, let me do it like this. I won't be able to destructure a otherwise.

1:25pm.

```
  ≤-begin_ : ∀ {x y : ℕ}
    → ∀ (A : Set)
    → {A ≡ x ≤ y}
    → {A}
      -----
    → A
```
```
———— Error —————————————————————————————————————————————————
E:\TheoremProving\plfa\Equality.lagda:378,7-18
{A ≡ x ≤ y} cannot appear by itself. It needs to be the argument to
a function expecting an implicit argument.
when scope checking {A ≡ x ≤ y}
```

No, this is not going for me.

1:30pm. I am just not sure how to do this. This might be a good question for the Agda SO.

Let me compose it then and I will move to the next thing.

1:40pm. No, I still want to think about it more. Now that I considered the thing more thoroughly, what is the type of `the`?

1:45pm. https://docs.racket-lang.org/pie/index.html#%28form._%28%28lib._pie%2Fmain..rkt%29._the%29%29

I am not sure where in the Agda library I can find it, so I am going to have to figure it out on my own.

1:45pm.

```
  the : (A : Set) → A → A
  the type expr = expr
```

This is probably what I am looking for...

2:35pm. Had to take a break from this in order to do the chores. Today it was different.

I was cleaning out those old Bug magazines that we had stored around. I must have stopped reading those around a decade ago, but nonetheless they were a part of my childhood. They were my dreams. Seeing them strewn around in the yard as rain fell down on them and then loading and hauling them only to throw them away was a bit sad. The wet paper, defenseless against the onslaught of time and the weather got thrown into the trash. Just where did the desire go?

I must continue working so that I find my power. And so that my life does not end up as that paper. I must do now what I could not in the time long past. Back then I believed in the future me. I cannot betray those expectations.

As the most ardent believer in computing, I have the responsibility to bring this era to an end.

Though I hope others are working towards the same end, it is not like just anybody can discover Calculus. Get rid of Newton and Leibnitz, and you would delay the discovery by quite a few years. To those living in the now, time is off the essence. Those looking for the answer are not replaceable.

2:40pm. Now let me resume. I've had time to think how I am going to do this exercise. Let me do it by chaining `the` like the above.

Let me do it from the top. I need a different design.

2:50pm.

```
  +-monoʳ-≤ : ∀ (n p q : ℕ) → p ≤ q → n + p ≤ n + q
  +-monoʳ-≤ zero    p q p≤q  =  p≤q
  +-monoʳ-≤ (suc n) p q =
    →-begin
      (p ≤ q)
    →⟨ fun p≤q => s≤s (+-monoʳ-≤ n p q p≤q) ⟩
      (n + p ≤ n + q)
    →∎
```

This kind of programming is exactly the kind I am good at. Won't be beat when it comes to design. Though given that to implement this I need lambdas which have not been covered yet, I am not sure what exactly to do as this juncture. This thing here is just a sketch.

2:55pm. Let me implement the functions and I'll move on the next exercise. I do not care about doing the proofs.

3:20pm.

```
_→⟨_⟩_ : ∀ {A : Set} →  A → (A → B) → ∀ (B : Set) → B
```

Ah, I am just not sure about this. That `∀ (B : Set)` needs to come in earlier, but then I would not be able to do the visual ordering properly.

...Forget this stuff. I am just not that interested in this exercise unless for some reason the others turn out to rely on it which should not be the case.

I'll figure it out as I go along.

3:35pm.

```
refl-≐ : ∀ {A : Set} {x : A}
  → x ≐ x
```

Had no idea it was possible to write out the parameters like this.

3:45pm. Uf...

```
_≐_ : ∀ {A : Set} (x y : A) → Set₁
_≐_ {A} x y = ∀ (P : A → Set) → P x → P y

refl-≐ : ∀ {A : Set} {x : A}
  → x ≐ x
refl-≐ P Px  =  Px

trans-≐ : ∀ {A : Set} {x y z : A}
  → x ≐ y
  → y ≐ z
    -----
  → x ≐ z
trans-≐ x≐y y≐z P Px  =  y≐z P (x≐y P Px)

sym-≐ : ∀ {A : Set} {x y : A}
  → x ≐ y
    -----
  → y ≐ x
sym-≐ {A} {x} {y} x≐y P  =  Qy
  where
    Q : A → Set
    Q z = P z → P x
    Qx : Q x
    Qx = refl-≐ P
    Qy : Q y
    Qy = x≐y Q Qx
```

I need to study this for a bit. But is this really equality?

3:35pm. Damn, this is complex. How does anybody even think of this? Just what do these people think about in their spare time?

Did Leibnitz really do this at the end of 17th century? If so, kudos to him.

```
    Qy : Q y
    Qy = x≐y Q Qx
```

This part is amazing. Usually I'd expect P to be passed everywhere, but if I inlined this I'd get...

```
Qy = x≐y (fun z => P z → P x) (refl-≐ P)
```

4pm. I am not really sure what the intermediate steps are here.

4:05pm. It must have been over a year since I could not do a functional programs in my head.

Very well, let me spend some time studying this. Somehow I might not finish the chapter today, but 3h should be enough.

```
_≐_ : ∀ {A : Set} (x y : A) → Set₁
_≐_ {A} x y = ∀ (P : A → Set) → P x → P y
```

First let me note what this thing is. Forget about the type for a moment. I need to focus on the terms.

`refl-≐ P` is just `fun Px => Px`. Let me think of it as the identity in this context.

`x≐y` already has `x` and `y` partially applied. Then...

```
∀ (P : A → Set) → P x → P y
```

If I applied P to it, it would just return `P x -> P y`.

One thing that makes it confusing is a forall is on the term level. I had no idea that was possible.

But I can just think of `P` as a regular function.

Suppose I replace it with Q.

That would end up being `(fun z => P z -> P x) x -> (fun z => P z -> P x) y`. Simplifying that for a step, I get...

```
(P x -> P x) -> P y -> P x
```

Hmmmm....then applying the identity as the first argument, I exactly get `P y -> P x`.

4:20pm. Amazingly it checks out.

So this is mathematics?

I couldn't think of this myself, but I can grok it.

4:55pm. Had to do a bit more chores. Let me resume. I am 90% done with the chapter by this point.

5pm. https://plfa.github.io/Isomorphism/

Done with the Equality chapter. The next one is this.

Once I am done with it, I will be done with half of the first book.

https://plfa.github.io/Statistics/

Eyeballing it, the two books look similar in length. I have my work cut out for me. After I am done with this I will return to Coq. I still want to conquer all the challenges in the original SF. My heart won't be at peace unless I do that.

Let me go through the next chapter for just a bit and then I will call it a day."