I feel like my focus wasn't what it could have been for the last month. But I am starting to grasp the feeling and the desire to get to the root of it all.
Everything I've done up to this point has just been the prelude.

It is not really about math - it is about my own intuition. It is about using machines to fix my own mental flaws. It is about making math just another tool in my toolbox, just like vanilla programming and using to analyze my programming at a deeper level.

It is ironic - looking up at something and doing nothing is the same as looking down on it. I've had a certain fear of really just diving into it, but now I've overcome it. If I continue on this path I will just keep getting better.

For a programmer, my ability at mathematics is just too low. It really was a huge mistake to start out with Logo and Basic back in elementary school. And for years now, I had a loss of control over my own center, but sometimes like now I'd get sparks of inspiration and see the hierarchy of goals and the truth of purpose.

I must not let it end in nothing. I must draw out its true power even if means killing it and putting it back together again.

That is the true essence of programming. It is one thing to do it to computer programs, but it applies to myself as well. It might in the end turn out to be the true essence of learning as well.

It has been a hard slog and it will continue to be hard, but as I look in myself for a moment I thought I saw the sparks of confidence to conquer the infinities - and bring this world to its knees.