"10am. Ah, I was the impression that I finished Isomorphism yesterday, but I only went halfway through the chapter. Let me finish reading these tweets and I will go through it properly.
10:10am. Focus me, stop reading the latest news on US vs China. Though I am surprised as to how much the US companies are willing to bend over to the government. I did not expect through would be willing to spread their ass cheeks at a moment's notice.

Let me learn those isomorphisms.

10:15am.

```
module ≃-Reasoning where

  infix  1 ≃-begin_
  infixr 2 _≃⟨_⟩_
  infix  3 _≃-∎

  ≃-begin_ : ∀ {A B : Set}
    → A ≃ B
      -----
    → A ≃ B
  ≃-begin A≃B = A≃B

  _≃⟨_⟩_ : ∀ (A : Set) {B C : Set}
    → A ≃ B
    → B ≃ C
      -----
    → A ≃ C
  A ≃⟨ A≃B ⟩ B≃C = ≃-trans A≃B B≃C

  _≃-∎ : ∀ (A : Set)
      -----
    → A ≃ A
  A ≃-∎ = ≃-refl
```

This clinches it. I was supposed to make `≤-Reasoning` in this form.

Though in that case, it does not feel like it would be particularly useful.

10:30am.

```
#### Exercise `≃-implies-≲`

Show that every isomorphism implies an embedding.
\begin{code}
postulate
  ≃-implies-≲ : ∀ {A B : Set}
    → A ≃ B
      -----
    → A ≲ B
\end{code}

\begin{code}
≃-implies-≲ A≃B = ?
\end{code}
```

There is a lot in this chapter, so let me get my bearings. I know I just need to access the fields of the first and strip out the last part.

10:35am.

```
\begin{code}
≃-implies-≲ : ∀ {A B : Set}
  → A ≃ B
    -----
  → A ≲ B
\end{code}

\begin{code}
≃-implies-≲ A≃B =
  record
    { from = from A≃B
    ; to = to A≃B
    ; from∘to = from∘to A≃B
    }
\end{code}
```

Well here it is.

10:45am. Now remember about those sparks and stop surfing /g/. As interesting as the latest China vs US happening is, I need to focus on math. Forget outside reality.

```
⇔-refl : ∀ {A} →  A ⇔ A
⇔-refl =
  record
    { from = λ x → x
    ; to = λ x → x
    }
```

Yeah, this works. Now let me do symmetry.

10:50am.

```
⇔-sym : ∀ {A B} → A ⇔ B → B ⇔ A
⇔-sym A⇔B =
  record
    { from = to A⇔B
    ; to = from A⇔B
    }
```

```
E:\TheoremProving\plfa\Isomorphism.lagda:468,12-20
Cannot resolve overloaded projection from because no matching
candidate found
when checking that the expression from A⇔B has type B → A
```

What do I do about this?

11am.

```
⇔-sym : ∀ {A B : Set} → A ⇔ B → A ⇔ B
⇔-sym A⇔B =
  record
    { from = {! !} --
    ; to = {!to A⇔B!}  -- to A⇔B
    }
```

```
E:\TheoremProving\plfa\Isomorphism.lagda:468,14-20
Cannot resolve overloaded projection to because no matching
candidate found
when checking that the expression to A⇔B has type A → B
```

Holy shit, what is going on here?

11:10am. Ah, it turns I need to open the record and give it a fixity.

11:15am.

```
infix 0 _⇔_
open _⇔_

⇔-refl : ∀ {A} →  A ⇔ A
⇔-refl =
  record
    { from = λ x → x
    ; to = λ x → x
    }

⇔-sym : ∀ {A B : Set} → A ⇔ B → (B ⇔ A)
⇔-sym A⇔B =
  record
    { from = to A⇔B
    ; to = from A⇔B
    }
```

Finally.

Now let me do the last one.

11:20am.

```
⇔-trans : ∀ {A B C : Set} → A ⇔ B → B ⇔ C → A ⇔ C
⇔-trans A⇔B B⇔C =
  record
    { from = λ x → from A⇔B (from B⇔C x)
    ; to = λ x → to B⇔C (to A⇔B x)
    }
```

I am mostly getting Agda practice here as doing this even without looking at past solutions is pretty trivial.

11:25am.

```
data Bin : Set where
  nil : Bin
  x0_ : Bin → Bin
  x1_ : Bin → Bin
\end{code}
And ask you to define the following functions

    to : ℕ → Bin
    from : Bin → ℕ

which satisfy the following property:

    from (to n) ≡ n

Using the above, establish that there is an embedding of `ℕ` into `Bin`.
\begin{code}
-- Your code goes here
\end{code}
```

The next is this thing. I already made `to` and `from` and `from (to n) ≡ n`, but the exercise here is so trivial that I do not feel like doing it...

I already have proofs of all the necessary elements in one of the previous chapters. Let me just more on to the next one.

12:15pm. Done with breakfast.

https://homepages.inf.ed.ac.uk/wadler/papers/propositions-as-types/propositions-as-types.pdf

I've been slowly reading this paper while I was at it. Today I feel like getting into the swing of things.

12:45pm. "We might be tempted to conclude that lambda calculus is universal, but first let’s ponder the suitability of the word ‘universal’. These days the multiple worlds interpretation of quantum physics is widely accepted. Scientists imagine that in different universes one might encounter different fundamental constants, such as the strength of gravity or the Planck constant. But easy as it may be to imagine a universe where gravity differs, it is difficult to conceive of a universe where fundamental rules of logic fail to apply. Natural deduction, and hence lambda calculus, should not only be known by aliens throughout our universe, but also throughout others. So we may conclude it would be a mistake to characterise lambda calculus as a universal language, because calling it universal would be too limiting."

This quote is great. I came to the same conclusion on my own so it is always a pleasure to see it restated somewhere else.

1:15pm. ...If there is a single alternative to differentiability + backpropagation...that would be reversible computation. I had that hint while working on differentiable plasticity.

But I'll store that card away until I can actually play it.

Rather than start reading Nagatoro, now that I am finally done with Wadler's paper let me resume the book."